Consolidate duplicate landing page components into shared Competitive components
Replace brand-specific Glean/ChatGptEnterprise hero, logo bar, what/why section,
and testimonial slider components with 4 shared components in the Competitive folder,
reducing ~525 lines of duplicated code.
